Bugfixes and added support for Opentype Fonts in TheDotFactory (toolchain).
GxEPD drivers - removed experimental code, lengthened reset pulse of 4.2" display (was causing problems  when rendering grey levels), altered grey levels of 2.7" display

Arabicligaturizer - increased unicode character size from 16 to 32 bits in some parts of the code, which was preventing unicode characters above the BMP from rendering correctly.

Bidi - fixed rendering problem which caused two newlines to be output if  text was the width of the page followed by <br>. Now one of the newlines is suppressed so a blank line is not rendered when only one carriage return is wanted.

Diskfont - Combined reads of diskfont fontcharinfo structures from the sd card into a single read. Commented/removed some backchannel debug output.

index.htm (lectionary web server config page) - fixed a bug which caused the contrast slider not to be set to the correct value if its value (stored in EEPROM) was 0.

Lectionary -changed Lectionary::ReadingsFromEnum to ReadingsFromEnum (no longer a class enum).

Network - reduced the wifi connect timeout from 60sec to 25sec

utf8string - added support for utf8 character lengths up to 6 bytes (was incomplete - only supported up to 4 byte utf8 characters before). Now all unicode character planes should be possible.

rcdisplay.ino - removed, merged with romancalendar.ino

romancalendar.ino - merged rcdisplay.ino code, removed Lectionary:: class  references wherever ReadingsFromEnum variabless are used.

FontDialog - a WPF/Opentype font dialog, now used with TheDotFactory (using wpf interop., since TheDotFactory is a legacy windows forms program).

TheDotFactory - now uses Opentype typefaces, and a WPF font chooser which supports them. This means that SimSun (Chinese font) and other fonts with charcters above the Basic Multilingual Plane can now be used on the Lectionary.

+++ libraries/RomanCalendar/ArabicLigaturizer.cpp
@@ -624,7 +624,7 @@
             int limit = start + length;
             for (int i = start; i < limit; ++i) {
                 String ch = GetUtf8CharByIndex(dest, i); //char ch = dest[i];
-                uint16_t uch = (uint16_t)codepointUtf8(ch);
+                uint32_t uch = (uint32_t)codepointUtf8(ch);
 				
 				switch (getBidiDirection(uch)) {
                 case BIDI_L:
